refactor: movve addplacecard handler to main
