+++ frontend/src/hooks/useModal.js
@@ -4,12 +4,14 @@ import { ModalContext } from "../contexts";
 //this should include handle signout
 export const useModal = () => {
   const {
+    isAddPlacePopupOpen,
     isEditAvatarPopupOpen,
     isEditProfilePopupOpen,
     isConfirmDeletePopupOpen,
     isToolTipOpen,
     isLoading,
     status,
+    setIsAddPlacePopupOpen,
     setIsEditAvatarPopupOpen,
     setIsEditProfilePopupOpen,
     setIsConfirmDeletePopupOpen,
@@ -19,12 +21,14 @@ export const useModal = () => {
   } = useContext(ModalContext);
 
   return {
+    isAddPlacePopupOpen,
     isEditAvatarPopupOpen,
     isEditProfilePopupOpen,
     isConfirmDeletePopupOpen,
     isToolTipOpen,
     isLoading,
     status,
+    setIsAddPlacePopupOpen,
     setIsEditAvatarPopupOpen,
     setIsEditProfilePopupOpen,
     setIsConfirmDeletePopupOpen,
